Fixed xipfs_buffer_flush.
Now the internal XIPFS buffer is aligned with respect to
FLASHPAGE_WRITE_BLOCK_ALIGNMENT, and the flush is performed with
aligned api flashpage_write_and_verify.
This circumvents a bug while flushing the buffer on the last page of
the flash memory.

Bonus : we write 4 times less and then we flush 4 times faster than
the unaligned version.